+++ Molten.Math/RandomNumber.cs
@@ -1,13 +1,8 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
-using System.Runtime;
-using System.Runtime.CompilerServices;
-using System.Globalization;
 using System.Diagnostics.Contracts;
 using System.Numerics;
-using Molten.DoublePrecision;
 
 namespace Molten
 {
@@ -185,7 +180,7 @@ namespace Molten
             Contract.EndContractBlock();
 
             long range = (long)maxValue - minValue;
-            if (range <= (long)Int32.MaxValue)
+            if (range <= int.MaxValue)
             {
                 return ((int)(Sample() * range) + minValue);
             }
@@ -204,7 +199,7 @@ namespace Molten
             Contract.EndContractBlock();
 
             long range = (long)maxValue - minValue;
-            if (range <= (long)Int16.MaxValue)
+            if (range <= short.MaxValue)
             {
                 return (short)((int)(Sample() * range) + minValue);
             }
