+++ Molten.Math/Vectors/Byte2.cs
@@ -1,12 +1,11 @@
-using System;
-using System.Runtime.InteropServices;
-using System.Runtime.CompilerServices;
 using System.Globalization;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
 
 namespace Molten
 {
-	///<summary>A <see cref = "byte"/> vector comprised of two components.</summary>
-	[StructLayout(LayoutKind.Sequential, Pack=1)]
+    ///<summary>A <see cref = "byte"/> vector comprised of two components.</summary>
+    [StructLayout(LayoutKind.Sequential, Pack=1)]
 	public partial struct Byte2 : IFormattable
 	{
 		///<summary>The X component.</summary>
@@ -697,7 +696,7 @@ namespace Molten
         /// <param name="right">Second <see cref="Byte2"/> source vector.</param>
         public static byte Dot(ref Byte2 left, ref Byte2 right)
         {
-			return (byte)(((byte)left.X * right.X) + ((byte)left.Y * right.Y));
+			return (byte)((left.X * right.X) + (left.Y * right.Y));
         }
 
 		/// <summary>
