+++ Molten.Math/Vectors/Byte3.cs
@@ -1,12 +1,11 @@
-using System;
-using System.Runtime.InteropServices;
-using System.Runtime.CompilerServices;
 using System.Globalization;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
 
 namespace Molten
 {
-	///<summary>A <see cref = "byte"/> vector comprised of three components.</summary>
-	[StructLayout(LayoutKind.Sequential, Pack=1)]
+    ///<summary>A <see cref = "byte"/> vector comprised of three components.</summary>
+    [StructLayout(LayoutKind.Sequential, Pack=1)]
 	public partial struct Byte3 : IFormattable
 	{
 		///<summary>The X component.</summary>
@@ -733,7 +732,7 @@ namespace Molten
         /// <param name="right">Second <see cref="Byte3"/> source vector.</param>
         public static byte Dot(ref Byte3 left, ref Byte3 right)
         {
-			return (byte)(((byte)left.X * right.X) + ((byte)left.Y * right.Y) + ((byte)left.Z * right.Z));
+			return (byte)((left.X * right.X) + (left.Y * right.Y) + (left.Z * right.Z));
         }
 
 		/// <summary>
