+++ Molten.Math/Vectors/Byte4.cs
@@ -1,12 +1,11 @@
-using System;
-using System.Runtime.InteropServices;
-using System.Runtime.CompilerServices;
 using System.Globalization;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
 
 namespace Molten
 {
-	///<summary>A <see cref = "byte"/> vector comprised of four components.</summary>
-	[StructLayout(LayoutKind.Sequential, Pack=1)]
+    ///<summary>A <see cref = "byte"/> vector comprised of four components.</summary>
+    [StructLayout(LayoutKind.Sequential, Pack=1)]
 	public partial struct Byte4 : IFormattable
 	{
 		///<summary>The X component.</summary>
@@ -771,7 +770,7 @@ namespace Molten
         /// <param name="right">Second <see cref="Byte4"/> source vector.</param>
         public static byte Dot(ref Byte4 left, ref Byte4 right)
         {
-			return (byte)(((byte)left.X * right.X) + ((byte)left.Y * right.Y) + ((byte)left.Z * right.Z) + ((byte)left.W * right.W));
+			return (byte)((left.X * right.X) + (left.Y * right.Y) + (left.Z * right.Z) + (left.W * right.W));
         }
 
 		/// <summary>
