+++ Molten.Math/Vectors/Vector2UL.i.cs
@@ -1,10 +1,7 @@
-using System.Runtime.InteropServices;
-using System;
-
 namespace Molten
 {
-	///<summary>A <see cref = "ulong"/> vector comprised of 2 components.</summary>
-	public partial struct Vector2UL
+    ///<summary>A <see cref = "ulong"/> vector comprised of 2 components.</summary>
+    public partial struct Vector2UL
 	{
 
 #region Operators - Cast
@@ -12,8 +9,8 @@ namespace Molten
 		{
 			return new Vector2D()
 			{
-				X = (double)value.X,
-				Y = (double)value.Y,
+				X = value.X,
+				Y = value.Y,
 			};
 		}
 #endregion
