+++ Molten.Math/Vectors/Vector2US.i.cs
@@ -1,10 +1,7 @@
-using System.Runtime.InteropServices;
-using System;
-
 namespace Molten
 {
-	///<summary>A <see cref = "ushort"/> vector comprised of 2 components.</summary>
-	public partial struct Vector2US
+    ///<summary>A <see cref = "ushort"/> vector comprised of 2 components.</summary>
+    public partial struct Vector2US
 	{
 
 #region Operators - Cast
@@ -12,8 +9,8 @@ namespace Molten
 		{
 			return new Vector2F()
 			{
-				X = (float)value.X,
-				Y = (float)value.Y,
+				X = value.X,
+				Y = value.Y,
 			};
 		}
 #endregion
